Switch theme to use context, implement graph theme
